Fixes
+ Fix offset for display in gutter
+ Fix error when switching modes
