+++ lib/NpmLibrary.js
@@ -1,6 +1,5 @@
 'use babel';
 
-import map from 'lodash/map';
 import reduce from 'lodash/reduce';
 import find from 'lodash/find';
 import debounce from 'lodash/debounce';
@@ -16,6 +15,12 @@ import {
 import {
   addBadge,
 } from '../components/badge';
+
+import {
+  getView,
+  clearMarkers,
+  addTextMarkerToLine,
+} from '../helpers/atom';
 
 import {
   NPM_LIBRARY_DESCRIPTION,
@@ -37,19 +42,6 @@ const mapDependencies = dependencies => reduce(dependencies, (prev, version, dep
 ]), []);
 
 /**
- * Removes all the markers from a given TextEditor
- * @param  {TextEditor} textEditor
- */
-const clearMarkers = (textEditor) => {
-  const markers = textEditor.findMarkers({
-    npmLibraryDescription: true,
-  });
-  map(markers, (marker) => {
-    marker.destroy();
-  });
-};
-
-/**
  * Converts dependencies from Object to Array
  * @param  {Object} contents Parsed `package.json` file
  * @return {Object}          Object containing `dependencies` as Array
@@ -62,76 +54,8 @@ const mapDependenciesFromContents = contents => ({
 const doesTextMatchDependency = text => dependency =>
   text.match(new RegExp(`"${dependency}":`));
 
-const getRangeFromLine = (textEditor, line) => {
-  const text = textEditor.lineTextForBufferRow(line);
-
-  const characters = reduce(text, (prev, char) => {
-    const currentChar = prev.currentChar + 1;
-
-    if (prev.isTabulation && /\s|\t/.test(char)) {
-      return {
-        ...prev,
-        currentChar,
-      };
-    }
-
-    return {
-      isTabulation: false,
-      currentChar,
-      textStartsAt: prev.isTabulation ? currentChar : prev.textStartsAt,
-    };
-  }, {
-    isTabulation: true, // Needed to know if the text of line is still in tabulation
-    currentChar: -1,
-    textStartsAt: 0,
-  });
-
-  return [[line, characters.textStartsAt], [line, characters.currentChar]];
-};
-
-const addTextMarkerToLine = ({ line, textEditor }) => {
-  const marker = textEditor.markBufferRange(
-    getRangeFromLine(textEditor, line),
-  );
-
-  marker.setProperties({
-    npmLibraryDescription: true,
-    line,
-  });
-
-  if (marker.isValid() && !marker.isDestroyed()) {
-    textEditor.decorateMarker(marker, {
-      type: 'text',
-      class: `${NPM_LIBRARY_DESCRIPTION} line-${line}`,
-    });
-  }
-};
-
-const calculatePixelsFromLine = (line, { textEditor, view }) => {
-  const lineHeight = textEditor.getLineHeightInPixels();
-  const lastVisibleRow = view.component.getLastVisibleRow();
-  const firstVisibleRow = view.component.getFirstVisibleRow();
-
-  const linesToBottom = lastVisibleRow - line;
-  const linesToTop = line - firstVisibleRow;
-
-  const pixelsToTop = lineHeight * linesToTop;
-  const pixelsToBottom = lineHeight * linesToBottom;
-
-  return {
-    lineHeight,
-    lastVisibleRow,
-    firstVisibleRow,
-    linesToBottom,
-    linesToTop,
-    pixelsToTop,
-    pixelsToBottom,
-  };
-};
-
 export default class NpmLibrary {
-  constructor(atom, validFiles = ['package.json'], preferences = {}) {
-    this.atom = atom;
+  constructor(validFiles = ['package.json'], preferences = {}) {
     this.validFiles = validFiles;
     this.dependencies = {};
     this.gutters = {};
@@ -144,10 +68,6 @@ export default class NpmLibrary {
     this.destroy = this.destroy.bind(this);
     this.onMouseMove = this.onMouseMove.bind(this);
     this.onDidChangeActiveTextEditor = this.onDidChangeActiveTextEditor.bind(this);
-  }
-
-  getView(textEditor) {
-    return this.atom.views.getView(textEditor);
   }
 
   /**
@@ -237,7 +157,7 @@ export default class NpmLibrary {
   }
 
   observeMouse(textEditor) {
-    const view = this.getView(textEditor);
+    const view = getView(textEditor);
     const {
       showTooltipWithMouseEvent,
     } = this.preferences;
@@ -286,7 +206,6 @@ export default class NpmLibrary {
 
       addTooltip({
         dependency,
-        ...calculatePixelsFromLine(line, { textEditor, view }),
         leftOffset: view.component.getGutterContainerWidth(),
       }, view);
     };
@@ -311,7 +230,7 @@ export default class NpmLibrary {
 
     const text = textEditor.getText();
     if (!isJSON.strict(text)) {
-      this.atom.notifications.addWarning(
+      atom.notifications.addWarning(
         `${NPM_LIBRARY_DESCRIPTION}: There was a problem reading ${fileName}`,
       );
       return;
@@ -334,7 +253,7 @@ export default class NpmLibrary {
     } = mapDependenciesFromContents(contents);
 
     const textEditorKey = getEditorKey(textEditor);
-    const view = this.getView(textEditor);
+    const view = getView(textEditor);
 
     const lines = textEditor.getLineCount();
     for (let line = 0; line < lines; line += 1) {
@@ -367,7 +286,6 @@ export default class NpmLibrary {
     line,
     textEditor,
     textEditorKey,
-    view,
   }) {
     const {
       showTooltipWithMouseEvent,
@@ -392,10 +310,8 @@ export default class NpmLibrary {
           type: 'gutter',
           class: `${NPM_LIBRARY_DESCRIPTION} line-${line}`,
           item: addBadge(
-            atom,
             {
               dependency,
-              ...calculatePixelsFromLine(line, { view, textEditor }),
             },
           ),
         });
@@ -475,6 +391,10 @@ export default class NpmLibrary {
     };
   }
 
+  getEventListener(textEditor) {
+    return this.eventListeners[getEditorKey(textEditor)] || {};
+  }
+
   onDidChangeActiveTextEditor(textEditor) {
     if (!textEditor || !this.isValidEditor(textEditor)) {
       return;
@@ -500,7 +420,7 @@ export default class NpmLibrary {
     const {
       element,
       eventListener,
-    } = this.eventListeners[getEditorKey(textEditor)];
+    } = this.getEventListener(textEditor);
 
     if (element) {
       element.removeEventListener('mouseover', eventListener);
