+++ lib/npm-library-description.js
@@ -41,7 +41,7 @@ the tooltip will be shown by hovering the mouse over the package text.
   },
 
   startLibrary(preferences) {
-    this.npmLibrary = new NpmLibrary(atom, ['package.json'], preferences);
+    this.npmLibrary = new NpmLibrary(['package.json'], preferences);
   },
 
   activate() {
